pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nwave_copilot"]

[tool.hatch.build.targets.wheel.force-include]
"nWave/agents" = "nWave/agents"
"nWave/skills" = "nWave/skills"
"nWave/tasks/nw" = "nWave/tasks/nw"
"des" = "des"

[project]
name = "nwave-copilot"
version = "0.1.0"
description = "GitHub Copilot CLI plugin for the nWave software development framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "Michele Brissoni", email = "michele.brissoni@brix.consulting" },
    { name = "Alessandro Digioia", email = "alessandro.digioia@brix.consulting" },
]
keywords = ["cli", "installer", "ai", "developer-tools", "copilot", "nwave"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Installation/Setup",
    "Typing :: Typed",
]
dependencies = [
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff==0.15.0",
]

[project.urls]
Homepage = "https://github.com/nwave-ai/nwave"
Repository = "https://github.com/nwave-ai/nwave"

[project.scripts]
nwave-copilot = "nwave_copilot.cli:main"
